+++ src/main/java/com/cedarsoftware/io/Resolver.java
@@ -212,8 +212,14 @@ public abstract class Resolver {
             return (T) rootObj.getTarget();
         } else {
             if (rootObj.getType() == null) {
-                // If there is no explicit type hint in the JSON, use the provided root type.
-                rootObj.setType(rootType);
+                if (rootType != null) {
+                    rootObj.setType(rootType);
+                } else if (rootObj.getTypeString() != null) {
+                    Class<?> clazz = stringToClass(rootObj.getTypeString());
+                    rootObj.setType(clazz);
+                } else {
+                    rootObj.setType(rootType);
+                }
             }
             Object instance = (rootObj.getTarget() == null ? createInstance(rootObj) : rootObj.getTarget());
             if (rootObj.isFinished) {
@@ -532,6 +538,20 @@ public abstract class Resolver {
             return ClassUtilities.getClassIfEnum(coercedClass);
         }
         return null;
+    }
+
+    private Class<?> stringToClass(String className) {
+        Class<?> clazz = ClassUtilities.forName(className, readOptions.getClassLoader());
+        if (clazz == null) {
+            if (readOptions.isFailOnUnknownType()) {
+                throw new JsonIoException("Unknown type (class) '" + className + "' not defined.");
+            }
+            clazz = readOptions.getUnknownTypeClass();
+            if (clazz == null) {
+                clazz = LinkedHashMap.class;
+            }
+        }
+        return clazz;
     }
 
     // Determine the factory type, considering enums and collections